pyproject.toml
[project]
name = "resume-forge"
version = "0.1.0"
description = "Privacy-first AI resume tailoring CLI tool"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer[all]",
    "openai",
    "chromadb",
    "langchain",
    "langchain-community",
    "langchain-openai",
    "sentence-transformers",
    "rich",
    "pyyaml",
    "pydantic-settings",
    "unstructured",
    "markdown"
]

[project.scripts]
resume-forge = "resume_forge.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["resume_forge"]
